utils.py
import os
from PyPDF2 import PdfReader
from langchain_community.tools import DuckDuckGoSearchRun
from langchain_chroma import Chroma
from langchain_huggingface import HuggingFaceEmbeddings
from langchain_core.documents import Document

# Initialize Embedding Model
# using a small, efficient model for local embeddings
embedding_function = HuggingFaceEmbeddings(model_name="all-MiniLM-L6-v2")

def read_pdf(uploaded_file):
    """
    Reads a PDF file and returns the text content.
    Args:
        uploaded_file: The uploaded PDF file object.
    Returns:
        str: extracted text.
    """
    try:
        pdf_reader = PdfReader(uploaded_file)
        text = ""
        for page in pdf_reader.pages:
            text += page.extract_text() or ""
        return text
    except Exception as e:
        return f"Error reading PDF: {e}"

def search_web(query):
    """
    Searches the web using DuckDuckGo.
    Args:
        query: Search query string.
    Returns:
        str: Search results.
    """
    try:
        search = DuckDuckGoSearchRun()
        return search.invoke(query)
    except Exception as e:
        return f"Error searching web: {e}"

def init_chroma(persist_directory="./chroma_db"):
    """
    Initializes ChromaDB collection and seeds it with dummy data if empty.
    Returns:
        Chroma: The Chroma vector store instance.
    """
    try:
        # Create a Chroma vector store
        vectorstore = Chroma(
            collection_name="winning_proposals",
            embedding_function=embedding_function,
            persist_directory=persist_directory
        )
        
        # Check if empty, if so, seed with dummy data
        # Note: In a real app, we would check collection.count(), but vectorstore abstraction hides it slightly.
        # We'll just try to add, if it duplicates it might be an issue but Chroma handles IDs.
        # For simplicity, we assume if we just created it and it might be reused. 
        # Actually, let's just return the vectorstore. We will use a separate function to seed if needed, 
        # or just rely on the user to understand this is a demo.
        # Let's add a simple seeder check.
        
        # Simple dummy proposal for RAG retrieval
        dummy_proposal_1 = """
        **Project:** E-Commerce Mobile App
        **Timeline:** 12 Weeks
        **Cost:** $45,000
        **Tech Stack:** React Native, Node.js, PostgreSQL
        **Breakdown:**
        - Planning: 2 Weeks
        - Design: 2 Weeks
        - Frontend: 4 Weeks
        - Backend: 3 Weeks
        - Testing: 1 Week
        """
        
        dummy_proposal_2 = """
        **Project:** Corporate AI Chatbot
        **Timeline:** 8 Weeks
        **Cost:** $30,000
        **Tech Stack:** Python, LangChain, OpenAI, React
        **Breakdown:**
        - Requirements: 1 Week
        - Core AI Dev: 3 Weeks
        - Integration: 2 Weeks
        - UI/UX: 1 Week
        - Deployment: 1 Week
        """

        # We can add a method to add documents if the collection is empty?
        # For this demo, let's just return the vector store.
        # The user's prompt implies we should store "10-15" proposals.
        # We will add a helper to seed if the db is fresh (implied by just doing it once).
        
        # But for robust "app" behavior, usually you separate seeding. 
        # I'll add a 'seed_db' function.
        
        return vectorstore
    except Exception as e:
        print(f"Error initializing Chroma: {e}")
        return None

def seed_database(vectorstore):
    """Seeds the database with sample proposals."""
    dummy_data = [
        Document(page_content="Project: High-Scale E-Commerce\nDuration: 16 Weeks\nCost: $60k\nStack: React, Django, AWS\nDetails: specific focus on scalability.", metadata={"type": "proposal"}),
        Document(page_content="Project: Internal HR Portal\nDuration: 10 Weeks\nCost: $35k\nStack: Angular, .NET Core, Azure\nDetails: focused on security and role-based access.", metadata={"type": "proposal"}),
        Document(page_content="Project: AI-Powered Customer Support Agent\nDuration: 6 Weeks\nCost: $25k\nStack: Python, Llama-3, Streamlit\nDetails: fast turnaround, focus on accuracy.", metadata={"type": "proposal"}),
    ]
    # Check if we can add (simple way: just add, Chroma handles dedupe if IDs provided, but we won't provide IDs so it might dupe. It's fine for a demo).
    vectorstore.add_documents(dummy_data)
