requirements.txt
langchain
langchain-groq
langchain-community
langgraph
streamlit
chromadb
pypdf2
duckduckgo-search
python-dotenv
langchain-huggingface
sentence-transformers
